Create default backend template.
Ok, so I believe how this is going to work is that we're going to wait got a
GET request to "/", then send our index.html file. The index.html file
contains a bundle.js (that was compiled by babel), which is our ENTIRE React
app. I'm 85% sure that's how client-side rendering works. I think.